Move testutils to test-utils folder
