+++ airbyte-webapp/src/views/Connector/ServiceForm/ServiceForm.test.tsx
@@ -3,9 +3,9 @@ import { getByTestId, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import React from "react";
 import selectEvent from "react-select-event";
+import { render } from "test-utils/testutils";
 
 import { AirbyteJSONSchema } from "core/jsonSchema";
-import { render } from "utils/testutils";
 import { ServiceForm } from "views/Connector/ServiceForm";
 
 import { DestinationDefinitionSpecificationRead } from "../../../core/request/AirbyteClient";
